Adding course working
